+++ src/main/java/me/drex/villagerconfig/data/TradeGroup.java
@@ -3,6 +3,7 @@ package me.drex.villagerconfig.data;
 import com.google.common.collect.Sets;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import me.drex.villagerconfig.util.loot.VCLootContextParams;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.npc.AbstractVillager;
 import net.minecraft.world.level.storage.loot.LootContext;
@@ -12,6 +13,7 @@ import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
 import net.minecraft.world.level.storage.loot.providers.number.NumberProvider;
 import net.minecraft.world.level.storage.loot.providers.number.NumberProviders;
 
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
@@ -33,8 +35,10 @@ public class TradeGroup {
 
     public List<BehaviorTrade> getTrades(AbstractVillager villager) {
         LootParams lootParams = new LootParams.Builder((ServerLevel) villager.level())
-            .withOptionalParameter(LootContextParams.THIS_ENTITY, villager)
-            .create(LootContextParamSets.PIGLIN_BARTER);
+            .withParameter(LootContextParams.ORIGIN, villager.position())
+            .withParameter(LootContextParams.THIS_ENTITY, villager)
+            .withParameter(VCLootContextParams.NUMBER_REFERENCE, Collections.emptyMap())
+            .create(VCLootContextParams.VILLAGER_LOOT_CONTEXT);
         LootContext lootContext = new LootContext.Builder(lootParams).create(Optional.empty());
 
         List<BehaviorTrade> applicableTrades = trades.stream().filter(behaviorTrade -> behaviorTrade.compositeCondition.test(lootContext)).toList();