+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.1.2] - 2024-07-29
 ### Changed
 - Removed `trade_enchantments` field from `enchant_randomly` loot function
 
@@ -70,7 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated to 1.19
 
 ## [3.0.1] - 2022-03-03
-## Added
+### Added
 - Trade validator
 - Behaviour pack support
 - Infinite trade setting